refactor(charts): extract DelayedTooltipContent to its own file
The tooltip's dwell timer and its over-the-stack geometry test are the one
non-obvious thing in the chart; keeping them inline made time-series-chart's
index read as tooltip plumbing rather than as a chart. Pure move.
